Ensure fluxctl install gets mandatory args
By using `fluxctl install --config-file`, it's possible to generate a
configuration for fluxd in which it doesn't need all it's command-line
flags. However, we do want to make sure some are present _either_ in
the config file, or in the command-line arguments.

This commit checks that the mandatory args are present. It's still
possible to produce an invalid configuration, of course, but this at
least preserves the property that `fluxctl install` will balk if it
doesn't see flags set.

I've factored the config struct into its own package, which is not
stricly necessary for this change (it's not even imported).
