+++ pkg/install/generated_templates.gogen.go
@@ -21,6 +21,13 @@ var templates = func() http.FileSystem {
 			name:    "/",
 			modTime: time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC),
 		},
+		"/#flux-deployment.yaml.tmpl#": &vfsgen۰CompressedFileInfo{
+			name:             "#flux-deployment.yaml.tmpl#",
+			modTime:          time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC),
+			uncompressedSize: 8346,
+
+			compressedContent: []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xd4\x59\xdd\x73\xdb\xc6\x11\x7f\xe7\x5f\xb1\xa5\x3b\x63\x69\x86\x00\xa5\x28\x1f\x1d\x24\xca\xd4\xb1\x65\x59\x75\xec\x68\x24\xbb\x9d\x3c\x45\x47\x60\x49\xdc\xf0\x70\x87\xde\x1e\xc8\x70\xd4\xfc\xef\x9d\xbd\xc3\xc7\x81\xa2\x64\xd7\xed\x4b\xf9\x90\x58\xc0\xde\x7e\x7f\xfc\x6e\x91\x24\xc9\x44\xd4\xf2\xef\x68\x49\x1a\x9d\x81\xa8\x6b\x9a\x6f\x4e\x27\x6b\xa9\x8b\x0c\x5e\x61\xad\xcc\xae\x42\xed\x26\x15\x3a\x51\x08\x27\xb2\x09\x80\x16\x15\x66\xb0\x54\xcd\xef\xf7\xf7\x20\x97\x90\xbe\x17\x15\x52\x2d\x72\x84\x3f\xfe\x68\xdf\xfb\x3f\x33\xb8\xbf\x1f\xbf\xbd\xbf\x07\xd4\x05\x93\x51\x8d\x39\x33\xb3\x58\x2b\x99\x0b\xca\xe0\x74\x02\x40\xa8\x30\x77\xc6\xf2\x1b\x80\x4a\xb8\xbc\xfc\x59\x2c\x50\x51\x78\x10\xcb\x66\x6a\x67\x85\xc3\xd5\x2e\xbc\x74\xbb\x1a\x33\xb8\xc1\xdc\xa2\x70\x38\x01\x70\x58\xd5\x4a\x38\x6c\x99\x45\x16\xf0\x4f\x68\x6d\x9c\x70\xd2\xe8\x9e\x39\x40\x6d\x4d\x85\xae\xc4\x86\x52\x69\xe6\xb5\xb1\x2e\x83\xe9\xd9\xc9\xd9\xe9\x14\x9e\x81\x43\xa5\x22\x0a\x70\x06\x28\xb7\xa2\x46\x98\x57\xe8\xac\xcc\x89\x8d\xab\x8d\xd4\xee\x39\x01\x1f\x4e\x5b\xc6\x6a\x64\xc3\x9e\x15\x00\x9d\x2f\xfc\x2b\x53\xe0\xed\xc8\x0b\xfc\x5b\xa0\x13\xe9\xba\x59\xa0\xd5\xe8\xd0\x2b\x67\x28\x03\x25\x75\xcb\x82\x5d\x67\x37\x32\xc7\x17\x79\x6e\x1a\xed\xde\x8f\x25\x00\x6c\x8c\x6a\x2a\xec\x75\x48\x5a\x1d\x56\xd2\x25\x6b\xdc\xf5\x82\x88\xdd\xe7\x06\xc1\xdd\x93\x81\x5f\xc2\x47\x0a\x9f\x19\x11\x55\x81\x4b\xd1\x28\xf7\xce\x14\x98\xc1\xc9\xd7\x27\x27\xf0\x0c\xb6\x25\x6a\xa8\x58\x1b\x2c\xc0\xa2\x28\x12\xa3\xd5\x6e\x06\x5b\x84\xad\xd1\xcf\x1d\x2c\x10\xc4\x42\x21\x3b\x32\x2f\x2b\x53\x4c\x5a\x86\xcf\xe0\x43\x29\x09\x24\x81\x00\x57\xd5\x4b\x82\x86\xb0\x80\xa5\xb1\xb0\x42\x8d\x56\x38\xa9\x57\x70\x7b\xfb\x06\xd6\xb8\xa3\x14\xae\x34\xbc\xfd\x0b\xc1\x8f\xe7\x70\x9a\x9e\x9e\xcc\x7a\x2e\x9d\xec\x60\x02\x81\xb0\x18\xeb\x41\x86\x55\xd1\x88\x05\x08\x20\xac\x05\x67\x53\xeb\x28\xd8\x62\xcf\x26\x17\x1a\xb6\x56\x3a\x56\x34\x3d\xec\xbf\x15\xea\xde\x19\x58\xd5\x6e\xf7\x4a\xda\xd8\x89\x15\x16\xb2\xa9\x32\x78\x87\x95\xb1\xbb\xd8\x4e\x84\xa5\x51\xca\x6c\xd9\xa2\x56\xb4\x24\x6f\x6a\x43\xfc\x4c\x40\xde\x90\x33\x95\x64\x0f\xac\xb5\xd9\xea\xdf\x4a\x43\x8e\x7a\x16\x4b\xa9\x70\x06\xdb\x52\xe6\x25\xec\x4c\x03\x5b\xa9\x54\x30\xca\x19\x28\x0c\x17\x28\x3f\xe6\x43\xfc\x0f\x0b\x66\xab\x59\xed\x9e\x81\xc5\xda\x80\x15\xae\x44\x0b\xae\x14\xba\x15\xbc\x92\xae\x6c\x16\x60\xf8\x21\x82\x92\x6b\x4c\xe1\x57\xd3\x3c\x57\x0a\x84\x22\xd3\x89\x18\x3b\x1b\xa4\x03\xa9\x9d\xf1\x67\x72\xa3\x9d\x90\x1a\xed\x0c\x16\xa8\xcc\x36\x85\x5b\x1c\xbc\x5a\x3a\x57\x53\x36\x9f\x17\x26\xa7\x94\x13\x2b\x2f\x38\xad\x51\xcf\xb9\x66\xc9\xcd\x57\x8d\x2c\x90\xe6\x0d\x61\x52\x5b\xb9\x11\x0e\x7d\xea\xb1\x21\x69\xe9\x2a\xd5\x73\xea\x62\x41\x54\x26\xb9\xd1\x4b\xb9\xea\x5f\x01\x84\x07\xef\x44\x9d\x45\x0f\xe3\x0a\x4c\xa2\x63\x5f\x1a\x17\x5f\x9a\xf3\xc0\x64\x48\xbf\x4f\xc6\x64\x2b\xa9\xe4\x27\xa5\xd8\x20\x08\x28\xe4\x72\x89\x96\xbb\x6d\xc7\xa1\xad\xaa\xa1\xa3\xfa\x10\x04\x76\x71\x10\xb8\x2b\x6d\x64\x81\x9d\xdb\x97\x72\x55\x89\x7a\x50\x44\xba\x12\x84\x06\xd4\xce\xee\xbc\x0d\x77\x81\xe8\x6e\x06\x42\x17\xd0\xe8\xdc\x54\xdc\xe6\xfd\xf9\x60\xed\x3b\x1f\x4e\xa1\x8b\x9e\x0b\xea\x8d\xe7\x20\x91\xda\x78\x3e\x88\x00\xbb\xe1\x0b\x22\x10\x1d\xfb\x64\x04\x7c\x27\x70\x06\x64\xc5\x0d\x16\x2e\xaf\x2f\x7d\x13\x80\x23\x36\x8b\xe4\x4a\x4b\x3d\x08\x67\xe3\x36\x68\xe5\x52\xe6\xbe\xd3\x43\xdd\xd8\xda\x10\xd2\xf1\x67\x38\xb2\xe7\x12\xda\x47\xf0\x22\x3b\x88\xe5\x7d\x86\xe3\x40\xd8\xd5\x50\xa6\x8f\x78\x6c\x55\xaf\xb8\x7f\x50\xe4\x9a\x71\x0b\x7e\xf6\x48\x13\x7e\x78\xee\x40\x13\xee\xdc\x79\x7f\x9f\xf8\x41\xfd\xd2\x3b\xf9\xb5\x54\xf8\xd2\x68\xc7\x7a\xfb\x81\xfd\x79\x1e\xef\x5c\x23\xda\x88\x36\x36\xf8\x94\x1d\xcf\x3a\x15\x23\x4e\xc3\xf9\x30\x8f\x0b\x56\x80\xbd\x94\xb4\xc5\x96\x70\xe3\x02\xf6\x1f\xfb\x4a\x49\x3d\x38\xdc\xd4\x9e\xb1\xa4\x4e\x66\x91\xc2\x55\xaf\xfe\x0b\x7a\xd9\x65\x14\x93\x10\x3a\xd6\xce\xd9\x06\x67\xa3\xb0\x8d\xb5\xe4\x61\x42\xd4\x54\xc1\x16\x9e\x3c\xd0\xb3\x99\x81\xe1\xee\xb7\x95\x84\xbe\x81\xf5\xa4\x3d\xaf\xee\xc8\xad\x8f\x43\x7a\xd0\xad\xb1\x5e\xbd\x5b\x93\x38\xcf\x5b\xc3\x83\x6b\xfa\xe9\xf0\xa0\x3e\xf6\xaa\x63\x54\x50\x2c\x92\x11\x54\xd2\x4b\x68\x95\x38\xd2\xc6\x1d\xd0\xe4\xf8\x3f\x52\xe5\x33\xa6\xff\x67\x68\xd3\x22\xbc\x49\x6f\x5e\x98\x02\x0f\xb0\x47\x84\x4e\xda\xac\xb1\xe8\x67\xb4\x36\x30\xcd\xc2\x14\x98\x82\xac\xc4\x0a\xa9\xcf\xb2\x14\x5e\x4b\x5d\xf8\x4c\xaa\x78\xa4\x59\xcc\x87\x8e\x19\xc6\x99\x42\x41\xc8\x83\xcb\x1f\xe5\x06\xc0\xe0\x16\x84\xeb\x67\x4e\xd9\x2c\xd2\xc2\xe4\x6b\xb4\x69\x6e\xaa\xb9\x9d\x87\xf9\xe3\xff\x37\x77\xa2\x2f\xdb\xae\x87\x30\x48\x65\x00\xcb\x52\x9d\x58\x01\x6b\x9a\xf6\x34\x5e\x4c\x06\x2d\x43\x69\x62\x6e\xd9\x69\x7a\xfa\x5d\x7a\x3a\xa6\xbd\x6e\x94\xba\x36\x4a\xe6\xbb\x0c\xae\x96\xef\x8d\xbb\xb6\x48\xb1\x15\x16\xc9\x34\x36\x47\x8a\x43\x61\xf1\x9f\x0d\x92\x1b\x3d\x03\xc8\xeb\x26\x83\x6f\x4e\xaa\xd1\xc3\xca\xc3\x8c\x0c\xbe\xfd\xfa\x9d\x1c\xb0\xad\xb1\xf1\xe1\x64\x88\xcc\xb5\xc7\xb9\x67\x27\x67\x8c\xda\xa4\x5e\x1a\x5b\xf9\xa2\x11\xaa\xa7\x56\x72\x83\x1a\x89\xae\xad\x59\x60\xac\x01\xbb\xf4\x72\x9c\x33\x41\x54\x60\x38\x7e\x2c\x5c\x99\xc1\x5c\xd4\x32\x78\x7a\xf3\xed\x5c\x16\xa8\x9d\x74\xbb\xb4\x6e\x16\x11\xad\xd4\xd2\x49\xa1\x5e\xa1\x12\xbb\x5b\x9e\x0d\x05\x65\xf0\x4d\x44\xe0\x64\x85\xa6\x71\x07\xde\x31\xc0\x93\xff\x1f\xaa\x46\x03\x63\x14\x98\xc3\xd0\x1c\x02\xc4\xba\x0e\x9a\xa1\xcb\xbd\x66\xc5\x9c\xa8\x0c\x1d\xca\x5f\x97\x40\x99\x76\xd6\xad\x38\x64\x20\x75\xc8\xb9\xe7\x14\xce\x10\x95\xf3\x51\x0d\x77\x3e\xfb\x45\xab\x5d\xe6\xdb\x28\x73\x63\xfc\xed\xa7\xe3\xa2\x05\x15\x5c\x52\x35\xda\xa5\xb1\x39\x32\xd3\x00\xb8\x19\x6f\x3f\xa6\x78\x8c\x89\xc7\xba\x6f\x84\x6d\x75\x0f\x64\x5f\xa6\x7e\x54\xa3\x57\x3a\x57\x8d\x9f\xda\x7c\x6d\x08\xe0\xaa\x9b\xe8\x01\x97\x7e\x02\x46\x77\x40\xfa\x7b\x3e\xba\x07\x71\x87\x61\x56\x60\xae\x84\xe5\xeb\xc2\xc2\x6c\xa2\x06\xf0\x04\x04\x0d\xa3\x39\x36\xde\x1a\xe3\xe6\x29\x51\xf9\xa8\x01\x42\x8f\xa4\x4e\x07\x78\x34\x0d\x92\x67\x1d\x49\xc4\x01\xf5\x46\x5a\xa3\x3d\x18\x09\x38\x6f\xfa\xf6\xe3\x4f\x17\x2f\x7f\x79\xff\xfa\xea\x72\x1a\xe0\xc7\x8c\xfd\x61\x36\x68\xed\x18\x2b\x46\x6c\xfc\xb0\x5f\xec\x02\x92\x73\xea\x90\x8d\x0f\x40\xde\x43\x1b\x87\xe4\x64\xe2\x47\x0d\x65\x50\xc1\xb7\xe5\x4e\x1a\xb7\xe8\x08\x06\xb7\xda\xf9\x98\x44\x2c\xf6\xc1\x74\x1c\x74\x8f\xa4\xbb\x6b\x9f\xd0\x20\x94\x43\xab\xf9\x5a\xf7\x40\xe3\xa5\x35\x55\x8f\x65\x2a\x06\x01\x82\x38\xdd\x5a\x44\xc7\x6e\x50\x26\x5f\xd3\xc3\x60\xa3\xde\x64\x07\xfc\x32\xb8\x7b\xe4\x97\x8d\x50\x0d\x3e\xf0\xc9\xa7\x92\x78\x3f\x07\x3a\xbc\xf7\x44\x06\x30\xdc\x1c\xc3\xcc\x27\x80\xe6\x23\x79\xc9\x54\x01\x59\x8f\xe8\xc6\xfd\x61\x50\xba\xdd\xff\x3c\x84\x95\x03\x20\xf8\x2c\xdc\xf1\x48\xf6\x44\x72\xf6\x51\xc5\x63\x55\xbf\x15\xda\xc3\x41\x6a\xea\x5a\xed\xe0\xcd\x87\x0f\xd7\xb0\x10\x24\x73\x10\x8d\x2b\x19\x8b\xfa\x2e\x2e\x54\x40\x14\xc3\x3d\x98\x19\x6e\xa4\xf0\x4e\xbf\xbb\xbc\xfa\xf0\xdb\x8b\x8f\x1f\xde\x7c\xbc\xbd\xb8\xb9\xf3\xae\xee\x1f\xbd\xbd\xf8\xf5\x6e\x54\x6c\x1b\x61\xa5\x58\x28\xa4\xee\x62\x18\x31\xa4\x11\x5c\x6c\x73\xe7\xb5\x35\xd5\x38\x7f\x02\xd9\x0d\x2e\xb3\x91\xd7\x47\x28\x90\x9b\x2a\x9b\x30\x38\x80\xe3\x9d\x8d\xfc\x11\x5c\x10\x76\x33\x58\x30\x0a\xc8\x45\x5e\x32\xf4\xd6\xa3\xba\xea\xaf\x93\xec\x29\xe6\x3e\x8b\xb8\x18\xdb\xde\x17\xa3\x03\xed\x6e\xc9\x1f\x9c\x79\x21\xb9\xd0\x9d\x8f\x5d\x89\x14\xe7\xe1\x70\x87\x70\x5b\xc3\x5a\x36\xec\x27\x5f\xed\x7e\x83\xe6\x43\x0b\xa5\xd9\xfa\xbd\x8f\xd1\x1a\x73\x1f\x32\xe9\xc6\x79\x9b\x24\xbd\x01\xfe\xd2\xcf\xc2\xcf\xfb\x47\x69\x7b\xd9\x49\x69\x93\xa7\xb9\x6a\xc8\xa1\x4d\x79\x78\xa8\xd8\x25\x1f\x29\xf4\xb9\xc1\x15\x2f\x03\xe9\xd5\xf5\xc8\x28\x6e\x79\x84\xce\xef\x95\xc6\x55\x35\xe8\xd0\xd1\x73\x76\x39\xcb\x94\x7e\xd3\x13\xe5\x79\xac\x71\x4b\x7d\x3e\x19\x21\x5c\x49\x50\x35\xe4\x37\x5f\xde\x7b\x12\x8b\x50\xca\x0b\x3f\x54\x3d\xbe\xf4\x0b\xaf\xa3\x6e\x8b\x74\x1c\xeb\xd2\x35\xb6\xd0\x02\x38\x81\xa3\xbd\xd7\x48\x11\x1e\x44\x61\xb8\x26\x85\xb4\xe7\x0f\x46\x6e\xac\xd6\x4d\x84\x6e\x87\xe0\x7d\xbc\xf9\x39\x2c\xe6\x84\x5e\x85\x77\x97\xd2\xf9\x65\x11\x49\x67\xec\xae\x1f\x15\xaf\x19\x95\x47\xec\x9e\xaa\x39\x4e\x9b\xc8\xf6\xb6\x64\x0e\x96\x53\x5c\x0b\x1d\x6e\xff\xf3\x51\x5c\x99\xc7\xd9\xf0\xf7\xdb\x8b\x5f\x8f\xff\x1a\x56\x56\x1e\xd2\x37\x84\x76\x3e\x28\x9b\xc6\x85\xce\xfe\xe1\x72\x6a\xac\x3a\xbf\xbf\x87\xf4\x52\x3a\x36\x76\xd4\xb2\x02\xc5\xc2\x0a\x9d\x97\x1d\xd1\x4f\xfe\xaf\xb0\xc5\xe6\xb6\x77\x29\x7d\xc7\xa2\x43\x27\x19\x3f\xf2\xb9\x5b\x9f\x29\xf4\x37\x23\x75\x74\x60\x3a\x9b\xb6\xcb\x70\x45\x18\x1f\x7f\xba\xa9\x59\xe4\xc4\xcb\xc3\xb6\xa1\x12\x5a\x2e\xf9\x3e\xc0\x35\x44\xb2\x40\x1b\xc2\xb1\x77\xab\xf2\xbb\x38\x43\x08\x8d\x2e\xd0\xee\xc5\xd8\xa2\x12\x4e\x6e\xd0\xc3\x5d\xea\x32\x70\x35\x8a\xf3\x5e\x4d\xf6\xc6\x51\xb3\x28\xa4\x3d\x9d\x85\xff\x7f\xd5\x6f\xf6\x07\xe7\xf8\xcd\xfd\x21\xe7\xf8\x75\x78\xe7\xd5\x8e\xea\x00\x83\x8f\x84\xf6\xd0\x79\x0e\x6e\x1f\x39\xa6\x81\xc3\xe7\x2f\x2a\x21\x0f\x2a\x80\xfc\xa2\xe3\xd0\x51\x0d\xdf\x26\x0e\x86\x03\xb9\x95\x6c\x0d\x3b\x14\xb5\x5f\x5b\xb3\x9f\x18\x2d\x48\xb7\xb7\x78\x8a\x7d\xd5\xce\xdd\x76\xaa\x9e\x3f\x31\x66\xbb\x13\x2d\x2f\x3e\x75\xfe\xc3\x1a\x77\x20\x8b\x1f\x9f\x82\x50\x91\x36\x7c\x54\xb8\xc6\xe2\x68\xeb\x75\x40\x86\x7f\xbd\x4b\x7a\x7a\x1a\xb5\xa9\xae\x4b\x83\x74\x50\x0a\xf2\xe3\xdf\x68\xb5\x03\x91\xe7\x48\xa1\x93\x97\x18\x16\xc7\x47\xdd\x8e\xf2\x6e\x29\x14\xe1\xdd\xf1\xa4\xdb\x8a\x5c\x4a\x77\xd3\xe2\x86\x43\x31\xe8\x98\x7a\xfa\x87\x75\x70\x98\xec\x40\x7c\xc8\xd9\x26\x77\x41\xdf\xad\x5f\x21\x30\xac\x6c\x1c\xd0\x4e\xe7\xb0\x30\x66\xbd\x46\xac\x39\xdb\x7b\x55\xa7\x2b\xe9\xa6\x33\xa8\x50\xb0\xa3\xb9\x91\x81\xf0\xf7\xfa\xb6\x00\x9a\x9a\x9c\x45\x51\xf5\x95\x70\xbc\xa7\x18\xb3\x4e\xc8\x09\x87\xe7\xdc\x58\x1e\xcd\x17\x8d\xbf\xbb\x2e\x69\xa2\x49\x27\x34\x4c\x3b\x19\xd3\x6e\x0e\x45\x4c\x8e\x30\x5d\xa5\x33\xf8\x07\x32\x9a\x7d\xa9\x4c\x53\x1c\xa7\x7e\xa9\xe6\xcc\x1a\xc3\x5e\x4c\x58\x27\xf3\x46\x09\xdb\x05\xa3\xe5\xb2\x3f\x42\x5b\xa9\xe7\x5b\xe2\xfe\x99\x33\xaf\x74\xcb\x7c\xd3\xad\xb1\x6b\xea\x2f\xb8\x7b\xc7\xbc\xa0\x73\xb1\xc8\x4f\xbf\x3a\x7b\xf8\xdf\xd8\xe0\x8b\x90\x7d\x5d\x37\xea\x3f\xd0\x18\xfd\x44\x6a\xbc\x6b\xa9\x2f\x07\xe2\xbd\x0c\xe9\xf8\x25\x03\xbf\x73\x8f\x3d\x1f\xcf\x96\x43\x47\xbc\xe0\x47\x52\xe7\x16\xed\xe6\xc0\xa7\x3b\xbe\x84\x0c\xc8\x87\x6b\xf4\xfb\x78\x04\x8b\x35\x8f\xaf\x90\x65\x84\x2e\xfa\x1e\xf8\x3c\xfa\xa4\x18\x7d\x1b\xe4\xe0\xf8\x55\xb5\xbf\x08\xa4\x23\x2b\x95\x24\x87\x3a\x69\x55\x38\xcf\xce\x4e\xce\x4e\x27\x3f\x84\x1f\xbc\xb9\x78\xf1\x6a\xf2\xaf\xf0\xe3\x80\xb3\x3a\x66\x09\xf8\xdd\xf2\xeb\x93\xe2\xbb\x65\x9a\xa6\xb0\x94\xbf\x37\xf5\x9f\xfa\xa4\x0b\xe0\xba\xdf\x1e\x71\x3a\xfb\x55\x94\x53\x20\x35\x39\xa1\x54\x64\xc9\x2b\x49\x3e\x72\x94\x0b\xed\xcb\xc0\x2c\xc1\xe2\x4a\x52\xd8\xf0\x1f\x75\x37\x90\xb0\x44\x2b\x84\x13\x0b\x41\x78\x3c\x39\x0f\xbf\xc9\xff\x80\xd5\x8f\xe1\xf7\x5f\x9a\xd4\x25\xd5\x4d\x90\xb8\x6b\xb5\xb9\xed\x94\xd9\xcb\xac\x56\xb1\x5d\x52\x04\xba\xa4\xd3\x3a\x4e\x93\x76\x86\xbc\x28\x0a\x19\x16\x61\x0c\x72\x5e\xf0\x05\x6b\x94\x76\xc3\xfb\x01\xe7\x8e\x2e\x2d\xd6\x83\xa7\x4f\xf0\x49\xfc\x67\xf2\xf8\xc9\x58\x93\xf6\x5f\x9d\xd0\x5f\xea\x56\xe4\xab\xf7\xb7\x1d\x68\xa5\x59\xbf\xee\x6e\x21\x2c\xe8\xc2\x38\xea\x56\xe9\x95\xd8\x79\x1f\xaa\xcd\xb0\x1f\xd7\xa4\x8c\x59\x37\x35\x48\xa2\x06\x09\x8c\x06\x32\x15\xc2\xdb\xfe\x33\x33\x73\x6f\x6a\x1a\xbe\x5a\x14\x9a\xba\xbd\xe5\xf4\xbd\xd1\x38\x8d\xdf\x84\x5b\x60\xfc\xdd\x22\x08\xa7\xf1\xa7\x8c\xee\x3a\xe8\xf5\x1b\xbd\xe9\xef\xca\xd3\xd3\xe9\xe4\xdf\x01\x00\x00\xff\xff\x40\xf6\xc6\xe6\x9a\x20\x00\x00"),
+		},
 		"/flux-account.yaml.tmpl": &vfsgen۰CompressedFileInfo{
 			name:             "flux-account.yaml.tmpl",
 			modTime:          time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC),
@@ -38,9 +45,9 @@ var templates = func() http.FileSystem {
 		"/flux-deployment.yaml.tmpl": &vfsgen۰CompressedFileInfo{
 			name:             "flux-deployment.yaml.tmpl",
 			modTime:          time.Date(1970, 1, 1, 0, 0, 0, 0, time.UTC),
-			uncompressedSize: 8042,
+			uncompressedSize: 8438,
 
-			compressedContent: []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xd4\x59\xdd\x73\xdb\x36\x12\x7f\xf7\x5f\xb1\xa3\xdc\x4c\xe2\x19\x91\xb2\xea\xb6\x77\xc3\x9e\x3a\x97\xe6\xc3\xf5\xa5\x49\x3d\x76\x72\x37\x7d\xaa\x21\x72\x25\x62\x04\x02\x3c\x2c\x28\x55\xe3\xe9\xff\x7e\xb3\x00\x3f\x40\x49\x76\x72\x79\xbb\x3c\xc4\x36\xb9\x58\x2c\x7e\xfb\xf5\xc3\x32\x49\x92\x33\x51\xcb\x7f\xa1\x25\x69\x74\x06\xa2\xae\x69\xb6\x9d\x9f\x6d\xa4\x2e\x32\x78\x8d\xb5\x32\xfb\x0a\xb5\x3b\xab\xd0\x89\x42\x38\x91\x9d\x01\x68\x51\x61\x06\x2b\xd5\xfc\xf1\xf0\x00\x72\x05\xe9\x07\x51\x21\xd5\x22\x47\xf8\xf3\xcf\xf6\xbd\xff\x33\x83\x87\x87\xf1\xdb\x87\x07\x40\x5d\xb0\x18\xd5\x98\xb3\x32\x8b\xb5\x92\xb9\xa0\x0c\xe6\x67\x00\x84\x0a\x73\x67\x2c\xbf\x01\xa8\x84\xcb\xcb\x5f\xc4\x12\x15\x85\x07\xf1\xde\x2c\xed\xac\x70\xb8\xde\x87\x97\x6e\x5f\x63\x06\xb7\x98\x5b\x14\x0e\xcf\x00\x1c\x56\xb5\x12\x0e\x5b\x65\xd1\x09\xf8\x9f\xd0\xda\x38\xe1\xa4\xd1\xbd\x72\x80\xda\x9a\x0a\x5d\x89\x0d\xa5\xd2\xcc\x6a\x63\x5d\x06\x93\xcb\x8b\xcb\xf9\x04\x9e\x81\x43\xa5\x22\x09\x70\x06\x28\xb7\xa2\x46\x98\x55\xe8\xac\xcc\x89\x0f\x57\x1b\xa9\xdd\x73\x02\x5e\x9c\xb6\x8a\xd5\xe8\x0c\x07\xa7\x00\xe8\xb0\xf0\xaf\x4c\x81\x77\x23\x14\xf8\xdf\x12\x9d\x48\x37\xcd\x12\xad\x46\x87\xde\x38\x43\x19\x28\xa9\x5b\x15\x0c\x9d\xdd\xca\x1c\x5f\xe6\xb9\x69\xb4\xfb\x30\xde\x01\x60\x6b\x54\x53\x61\x6f\x43\xd2\xda\xb0\x96\x2e\xd9\xe0\xbe\xdf\x88\x18\x3e\x37\x6c\xdc\x3d\x19\xf4\x25\xbc\xa4\xf0\x91\x11\x49\x15\xb8\x12\x8d\x72\xef\x4d\x81\x19\x5c\x7c\x7b\x71\x01\xcf\x60\x57\xa2\x86\x8a\xad\xc1\x02\x2c\x8a\x22\x31\x5a\xed\xa7\xb0\x43\xd8\x19\xfd\xdc\xc1\x12\x41\x2c\x15\x32\x90\x79\x59\x99\xe2\xac\x55\xf8\x0c\x3e\x96\x92\x40\x12\x08\x70\x55\xbd\x22\x68\x08\x0b\x58\x19\x0b\x6b\xd4\x68\x85\x93\x7a\x0d\x77\x77\x3f\xc3\x06\xf7\x94\xc2\xb5\x86\x77\x7f\x23\xf8\x71\x01\xf3\x74\x7e\x31\xed\xb5\x74\x7b\x87\x23\x10\x08\x8b\xb1\x1d\x64\xd8\x14\x8d\x58\x80\x00\xc2\x5a\x70\x34\xb5\x40\xc1\x0e\x7b\x35\xb9\xd0\xb0\xb3\xd2\xb1\xa1\xe9\x69\xfc\xd6\xa8\x7b\x30\xb0\xaa\xdd\xfe\xb5\xb4\x31\x88\x15\x16\xb2\xa9\x32\x78\x8f\x95\xb1\xfb\xf8\x9c\x08\x2b\xa3\x94\xd9\xf1\x89\xda\xad\x25\xf9\xa3\x36\xc4\xcf\x04\xe4\x0d\x39\x53\x49\x46\x60\xa3\xcd\x4e\xff\x5e\x1a\x72\xd4\xab\x58\x49\x85\x53\xd8\x95\x32\x2f\x61\x6f\x1a\xd8\x49\xa5\xc2\xa1\x9c\x81\xc2\x70\x82\xf2\x63\x5e\xc4\xbf\x58\x30\x3b\xcd\x66\xf7\x0a\x2c\xd6\x06\xac\x70\x25\x5a\x70\xa5\xd0\xed\xc6\x6b\xe9\xca\x66\x09\x86\x1f\x22\x28\xb9\xc1\x14\x7e\x33\xcd\x73\xa5\x40\x28\x32\xdd\x16\x63\xb0\x41\x3a\x90\xda\x19\xbf\x26\x37\xda\x09\xa9\xd1\x4e\x61\x89\xca\xec\x52\xb8\xc3\x01\xd5\xd2\xb9\x9a\xb2\xd9\xac\x30\x39\xa5\x1c\x58\x79\xc1\x61\x8d\x7a\xc6\x39\x4b\x6e\xb6\x6e\x64\x81\x34\x6b\x08\x93\xda\xca\xad\x70\xe8\x43\x8f\x0f\x92\x96\xae\x52\xbd\xa6\xce\x17\x44\x65\x92\x1b\xbd\x92\xeb\xfe\x15\x40\x78\xf0\x5e\xd4\x59\xf4\x30\xce\xc0\x24\x5a\xf6\xb5\x7e\xf1\xa9\x39\x0b\x4a\x86\xf0\xfb\xac\x4f\x76\x92\x4a\x7e\x52\x8a\x2d\x82\x80\x42\xae\x56\x68\xb9\xda\x76\x1a\xda\xac\x1a\x2a\xaa\x77\x41\x50\x17\x3b\x81\xab\xd2\x56\x16\xd8\xc1\xbe\x92\xeb\x4a\xd4\x83\x21\xd2\x95\x20\x34\xa0\x76\x76\xef\xcf\x70\x1f\x84\xee\xa7\x20\x74\x01\x8d\xce\x4d\xc5\x65\xde\xaf\x0f\xa7\x7d\xef\xdd\x29\x74\xd1\x6b\x41\xbd\xf5\x1a\x24\x52\xeb\xcf\x23\x0f\x30\x0c\x5f\xe1\x81\x68\xd9\x67\x3d\xe0\x2b\x81\x33\x20\x2b\x2e\xb0\x70\x75\x73\xe5\x8b\x00\xbc\xe0\x63\x91\x5c\x6b\xa9\x87\xcd\xf9\x70\x5b\xb4\x72\x25\x73\x5f\xe9\xa1\x6e\x6c\x6d\x08\xe9\xfc\x0b\x80\xec\xb5\x84\xf2\x11\x50\x64\x80\x78\xbf\x2f\x00\x0e\x84\x5d\x0f\x69\xfa\x08\x62\xeb\x7a\xcd\xf5\x83\x22\x68\xc6\x25\xf8\xd9\x23\x45\xf8\x78\xdd\x89\x22\xdc\xc1\xf9\xf0\x90\xf8\x46\xfd\xca\x83\xfc\x56\x2a\x7c\x65\xb4\x63\xbb\x7d\xc3\xfe\x32\xc4\x3b\x68\x44\xeb\xd1\xc6\x06\x4c\x19\x78\xb6\xa9\x18\x69\x1a\xd6\x87\x7e\x5c\xb0\x01\x8c\x52\xd2\x26\x5b\xc2\x85\x0b\x18\x3f\xc6\x4a\x49\x3d\x00\x6e\x6a\xaf\x58\x52\xb7\x67\x91\xc2\x75\x6f\xfe\x4b\x7a\xd5\x45\x14\x8b\x10\x3a\xb6\xce\xd9\x06\xa7\x23\xb7\x8d\xad\xe4\x66\x42\xd4\x54\xe1\x2c\xdc\x79\xa0\x57\x33\x05\xc3\xd5\x6f\x27\x09\x7d\x01\xeb\x45\x7b\x5d\xdd\x92\x3b\xef\x87\xf4\x24\xac\xb1\x5d\x3d\xac\x49\x1c\xe7\xed\xc1\x03\x34\x7d\x77\x38\xca\x8f\x83\xec\x18\x25\x14\x6f\xc9\x0c\x2a\xe9\x77\x68\x8d\x78\xa1\x8d\x3b\x61\xc9\xf9\xff\x64\xca\x17\x74\xff\x2f\xb0\xa6\x65\x78\x67\xfd\xf1\x42\x17\x38\xe2\x1e\x11\x3b\x69\xa3\xc6\xa2\xef\xd1\xda\xc0\x24\x0b\x5d\x60\x02\xb2\x12\x6b\xa4\x3e\xca\x52\x78\x2b\x75\xe1\x23\xa9\xe2\x96\x66\x31\x1f\x2a\x66\x68\x67\x0a\x05\x21\x37\x2e\xbf\x94\x0b\x00\x93\x5b\x10\xae\xef\x39\x65\xb3\x4c\x0b\x93\x6f\xd0\xa6\xb9\xa9\x66\x76\x16\xfa\x8f\xff\x31\x73\xa2\x4f\xdb\xae\x86\x30\x49\x65\x02\xcb\xbb\x3a\xb1\x06\xb6\x34\xed\x65\xfc\x36\x19\xb4\x0a\xa5\x89\xb5\x65\xf3\x74\xfe\xd7\x74\x3e\x96\xbd\x69\x94\xba\x31\x4a\xe6\xfb\x0c\xae\x57\x1f\x8c\xbb\xb1\x48\xf1\x29\x2c\x92\x69\x6c\x8e\x14\xbb\xc2\xe2\x7f\x1a\x24\x37\x7a\x06\x90\xd7\x4d\x06\xdf\x5d\x54\xa3\x87\x95\xa7\x19\x19\x7c\xff\xed\x7b\x39\x70\x5b\x63\xe3\xc5\xc9\xe0\x99\x1b\xcf\x73\x2f\x2f\x2e\x99\xb5\x49\xbd\x32\xb6\xf2\x49\x23\x54\x2f\xad\xe4\x16\x35\x12\xdd\x58\xb3\xc4\xd8\x02\x86\xf4\x6a\x1c\x33\x61\xab\xa0\x70\xfc\x58\xb8\x32\x83\x99\xa8\x65\x40\x7a\xfb\xfd\x4c\x16\xa8\x9d\x74\xfb\xb4\x6e\x96\x91\xac\xd4\xd2\x49\xa1\x5e\xa3\x12\xfb\x3b\xee\x0d\x05\x65\xf0\x5d\x24\xe0\x64\x85\xa6\x71\x27\xde\x31\xc1\x93\xff\x1f\xa6\x46\x0d\x63\xe4\x98\xd3\xd4\x1c\x02\xc5\xba\x09\x96\xa1\xcb\xbd\x65\xc5\x8c\xa8\x0c\x15\xca\x5f\x97\x40\x99\xb6\xd7\xad\xd9\x65\x20\x75\x88\xb9\xe7\x14\xd6\x10\x95\xb3\x51\x0e\x77\x98\xfd\xaa\xd5\x3e\xf3\x65\x94\xb5\x31\xff\xf6\xdd\x71\xd9\x92\x0a\x4e\xa9\x1a\xed\xca\xd8\x1c\x59\x69\x20\xdc\xcc\xb7\x1f\x33\x3c\xe6\xc4\x63\xdb\xb7\xc2\xb6\xb6\x07\xb1\xaf\x33\x3f\xca\xd1\x6b\x9d\xab\xc6\x77\x6d\xbe\x36\x04\x72\xd5\x75\xf4\xc0\x4b\x3f\x43\xa3\x3b\x22\xfd\x03\x2f\x3d\xa0\xb8\x43\x33\x2b\x30\x57\xc2\xf2\x75\x61\x69\xb6\x51\x01\x78\x82\x82\x86\xd6\x1c\x1f\xde\x1a\xe3\x66\x29\x51\xf9\xe8\x01\x84\x1e\xed\x3a\x19\xe8\xd1\x24\xec\x3c\xed\x44\x22\x0d\xa8\xb7\xd2\x1a\xed\xc9\x48\xe0\x79\x93\x77\x9f\x7e\x7a\xf3\xea\xd7\x0f\x6f\xaf\xaf\x26\x81\x7e\x4c\x19\x0f\xb3\x45\x6b\xc7\x5c\x31\x52\xe3\x9b\xfd\x72\x1f\x98\x9c\x53\xa7\xce\x78\x44\xf2\x8e\xcf\x38\x04\x27\x0b\x3f\x7a\x50\x26\x15\x7c\x5b\xee\x76\xe3\x12\x1d\xd1\xe0\xd6\x3a\xef\x93\x48\xc5\x21\x99\x8e\x9d\xee\x99\x74\x77\xed\x13\x1a\x84\x72\x68\x35\x5f\xeb\x8e\x2c\x5e\x59\x53\xf5\x5c\xa6\x62\x12\x20\x88\xc3\xad\x65\x74\x0c\x83\x32\xf9\x86\x8e\x9d\x8d\x7a\x9b\x9d\xc0\x65\x80\x7b\x84\xcb\x56\xa8\x06\x8f\x30\xf9\x5c\x10\x1f\xc6\x40\xc7\xf7\x9e\x88\x00\xa6\x9b\x63\x9a\xf9\x04\xd1\x7c\x24\x2e\x59\x2a\x30\xeb\x91\xdc\xb8\x3e\x0c\x46\xb7\xf3\x9f\x63\x5a\x39\x10\x82\x2f\xe2\x1d\x8f\x44\x4f\xb4\xcf\x21\xab\x78\x2c\xeb\x77\x42\x7b\x3a\x48\x4d\x5d\xab\x3d\xfc\xfc\xf1\xe3\x0d\x2c\x05\xc9\x1c\x44\xe3\x4a\xe6\xa2\xbe\x8a\x0b\x15\x18\xc5\x70\x0f\x66\x85\x5b\x29\x3c\xe8\xf7\x57\xd7\x1f\x7f\x7f\xf9\xe9\xe3\xcf\x9f\xee\xde\xdc\xde\x7b\xa8\xfb\x47\xef\xde\xfc\x76\x3f\x4a\xb6\xad\xb0\x52\x2c\x15\x52\x77\x31\x8c\x14\xd2\x88\x2e\xb6\xb1\xf3\xd6\x9a\x6a\x1c\x3f\x41\xec\x16\x57\xd9\x08\xf5\x11\x0b\xe4\xa2\xca\x47\x18\x00\x60\x7f\x67\x23\x3c\x02\x04\x61\x36\x83\x05\xb3\x80\x5c\xe4\x25\x53\x6f\x3d\xca\xab\xfe\x3a\xc9\x48\xb1\xf6\x69\xa4\xc5\xd8\xf6\xbe\x18\x2d\x68\x67\x4b\x7e\xe1\xd4\x6f\x92\x0b\xdd\x61\xec\x4a\xa4\x38\x0e\x87\x3b\x84\xdb\x19\xb6\xb2\x61\x9c\x7c\xb6\xfb\x09\x9a\x77\x2d\x94\x66\xe7\xe7\x3e\x46\x6b\xcc\xbd\xcb\xa4\x1b\xc7\x6d\x92\xf4\x07\xf0\x97\x7e\xde\x7c\xd1\x3f\x4a\xdb\xcb\x4e\x4a\xdb\x3c\xcd\x55\x43\x0e\x6d\xca\xcd\x43\xc5\x90\x7c\xa2\x50\xe7\x06\x28\x5e\x05\xd1\xeb\x9b\xd1\xa1\xb8\xe4\x11\x3a\x3f\x57\x1a\x67\xd5\x60\x43\x27\xcf\xd1\xe5\x2c\x4b\xfa\x49\x4f\x14\xe7\xb1\xc5\xad\xf4\xe2\x6c\xc4\x70\x25\x41\xd5\x90\x9f\x7c\x79\xf4\x24\x16\x21\x95\x97\xbe\xa9\x7a\x7e\xe9\x07\x5e\x2f\xba\x29\xd2\x79\x6c\x4b\x57\xd8\x42\x09\xe0\x00\x8e\xe6\x5e\x23\x43\xb8\x11\x85\xe6\x9a\x14\xd2\x2e\x8e\x5a\x6e\x6c\xd6\x6d\xc4\x6e\x07\xe7\x7d\xba\xfd\x25\x0c\xe6\x84\x5e\x87\x77\x57\xd2\xf9\x61\x11\x49\x67\xec\xbe\x6f\x15\x6f\x99\x95\x47\xea\x9e\xca\x39\x0e\x9b\xe8\xec\x6d\xca\x9c\x4c\xa7\x38\x17\x3a\xde\xfe\x97\x17\x71\x66\x9e\x67\xc3\xdf\xef\xde\xfc\x76\xfe\x8f\x30\xb2\xf2\x94\xbe\x21\xb4\xb3\xc1\xd8\x34\x4e\x74\xc6\x87\xd3\xa9\xb1\x6a\xf1\xf0\x00\xe9\x95\x74\x7c\xd8\x51\xc9\x0a\x12\x4b\x2b\x74\x5e\x76\x42\x3f\xf9\xbf\xc2\x14\x9b\xcb\xde\x95\xf4\x15\x8b\x4e\xad\x64\xfe\xc8\xeb\xee\x7c\xa4\xd0\x3f\x8d\xd4\xd1\x82\xc9\x74\xd2\x0e\xc3\x15\x61\xbc\xfc\xe9\xa2\x66\x91\x03\x2f\x0f\xd3\x86\x4a\x68\xb9\xe2\xfb\x00\xe7\x10\xc9\x02\x6d\x70\xc7\xc1\xad\xca\xcf\xe2\x0c\x21\x34\xba\x40\x7b\xe0\x63\x8b\x4a\x38\xb9\x45\x4f\x77\xa9\x8b\xc0\xf5\xc8\xcf\x07\x39\xd9\x1f\x8e\x9a\x65\x21\xed\x7c\x1a\x7e\x7e\xd3\x4f\xf6\x07\x70\xfc\xe4\xfe\x14\x38\x7e\x1c\xde\xa1\xda\x49\x9d\x50\xf0\x89\xd0\x9e\x5a\xcf\xce\xed\x3d\xc7\x32\x70\x7a\xfd\x9b\x4a\xc8\x93\x06\x20\xbf\xe8\x34\x74\x52\xc3\xb7\x89\x93\xee\x40\x2e\x25\x3b\xc3\x80\xa2\xf6\x63\x6b\xc6\x89\xd9\x82\x74\x07\x83\xa7\x18\xab\xb6\xef\xb6\x5d\x75\xf1\x44\x9b\xed\x56\xb4\xba\x78\xd5\xe2\xef\x1b\xdc\x83\x2c\x7e\x7c\x8a\x42\x45\xd6\xf0\x52\xe1\x1a\x8b\xa3\xa9\xd7\x89\x3d\xfc\xeb\x7d\xd2\xcb\xd3\xa8\x4c\x75\x55\x1a\xa4\x83\x52\x90\x6f\xff\x46\xab\x3d\x88\x3c\x47\x0a\x95\xbc\xc4\x30\x38\x7e\xd1\xcd\x28\xef\x57\x42\x11\xde\x9f\x9f\x75\x53\x91\x2b\xe9\x6e\x5b\xde\x70\xca\x07\x9d\x52\x2f\x7f\x9c\x07\xa7\xc5\x4e\xf8\x87\x9c\x6d\x72\x17\xec\xdd\xf9\x11\x02\xd3\xca\xc6\x01\xed\x75\x0e\x4b\x63\x36\x1b\xc4\x9a\xa3\xbd\x37\x75\xb2\x96\x6e\x32\x85\x0a\x05\x03\xcd\x85\x0c\x84\xbf\xd7\xb7\x09\xd0\xd4\xe4\x2c\x8a\xaa\xcf\x84\xf3\x03\xc3\x58\x75\x42\x4e\x38\x5c\x70\x61\x79\x34\x5e\x34\xfe\xe1\xba\xa0\x89\x3a\x9d\xd0\x30\xe9\xf6\x98\x74\x7d\x28\x52\xf2\x02\xd3\x75\x3a\x85\x7f\x23\xb3\xd9\x57\xca\x34\xc5\x79\xea\x87\x6a\xce\x6c\x30\xcc\xc5\x84\x75\x32\x6f\x94\xb0\x9d\x33\x5a\x2d\x87\x2d\xb4\xdd\x75\xb1\x23\xae\x9f\x39\xeb\x4a\x77\xac\x37\xdd\x19\xbb\xa1\xfe\x82\x7b\xb0\xcc\x6f\xb4\x10\xcb\x7c\xfe\xcd\xe5\xf1\xff\xf1\x81\xdf\x84\xe8\xeb\xaa\x51\xff\x81\xc6\xe8\x27\x42\xe3\x7d\x2b\x7d\x35\x08\x1f\x44\x48\xa7\x2f\x19\xf4\x2d\x3c\xf7\x7c\x3c\x5a\x4e\x2d\xf1\x1b\x3f\x12\x3a\x77\x68\xb7\x27\x3e\xdd\xf1\x25\x64\x60\x3e\x9c\xa3\x3f\xc4\x2d\x58\x6c\xb8\x7d\x85\x28\x23\x74\xd1\xf7\xc0\xe7\xd1\x27\xc5\xe8\xdb\x20\x3b\xc7\x8f\xaa\xfd\x45\x20\x1d\x9d\x52\x49\x72\xa8\x93\xd6\x84\x45\x76\x79\x71\x39\xef\x41\xba\xc5\xb5\x24\x67\xf7\xaf\x25\x31\xc4\x77\xb9\xd0\x3e\x5c\x0f\x90\xb2\xad\x58\x52\x04\xb9\x84\x5a\xc1\xf8\xd8\x6d\x4d\x7c\x59\x14\x32\x0c\x76\xb8\x69\xbf\xe4\x0b\xc3\x08\xc6\xe1\xfd\xc0\xdb\x46\x24\xdc\x7a\x32\xf0\x19\x3d\x89\xff\xec\x1b\x3f\x19\x5b\xd2\xfe\xd6\x6d\xfa\x6b\xdd\x6e\xf9\xfa\xc3\x5d\x47\xc2\x68\xda\x8f\x6f\x5b\x4a\x06\xba\x30\x8e\xba\xd1\x70\x25\xf6\x7e\x48\xa6\xb6\xc3\xbc\x57\x93\x32\x66\xd3\xd4\x20\x89\x1a\x24\x30\x1a\xc8\x54\x08\xef\xfa\xcf\xa6\xac\xbd\xa9\x69\x98\xc2\x17\x9a\xba\x39\xdc\xe4\x83\xd1\x38\x89\xdf\x84\x5b\x4d\x3c\x87\x0f\x9b\xd3\x78\x34\xdf\x5d\x6f\xbc\x7d\xa3\x37\xfd\xdd\x6f\x32\x9f\x9c\xfd\x37\x00\x00\xff\xff\x6b\x33\x36\xc3\x6a\x1f\x00\x00"),
+			compressedContent: []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xd4\x59\xdd\x73\xdb\x36\x12\x7f\xf7\x5f\xb1\xa3\xdc\x4c\xec\x19\x91\xb2\xeb\xb6\x77\xc3\x9e\x3b\x97\xe6\xc3\xf5\xa5\x49\x3d\x76\x72\x37\x7d\xaa\x21\x72\x25\x62\x04\x02\x3c\x2c\x28\x55\xe3\xe9\xff\x7e\xb3\x00\x3f\x40\x49\x96\x73\x79\x3b\x3f\xc4\x31\x09\x2c\x76\x7f\xfb\xf5\xc3\x32\x49\x92\x13\x51\xcb\x7f\xa1\x25\x69\x74\x06\xa2\xae\x69\xb6\xbe\x38\x59\x49\x5d\x64\xf0\x06\x6b\x65\xb6\x15\x6a\x77\x52\xa1\x13\x85\x70\x22\x3b\x01\xd0\xa2\xc2\x0c\x16\xaa\xf9\xe3\xf1\x11\xe4\x02\xd2\x8f\xa2\x42\xaa\x45\x8e\xf0\xe7\x9f\xed\x7b\xff\x67\x06\x8f\x8f\xe3\xb7\x8f\x8f\x80\xba\xe0\x65\x54\x63\xce\xc2\x2c\xd6\x4a\xe6\x82\x32\xb8\x38\x01\x20\x54\x98\x3b\x63\xf9\x0d\x40\x25\x5c\x5e\xfe\x22\xe6\xa8\x28\x3c\x88\xcf\xe6\xd5\xce\x0a\x87\xcb\x6d\x78\xe9\xb6\x35\x66\x70\x87\xb9\x45\xe1\xf0\x04\xc0\x61\x55\x2b\xe1\xb0\x15\x16\x59\xc0\x3f\x42\x6b\xe3\x84\x93\x46\xf7\xc2\x01\x6a\x6b\x2a\x74\x25\x36\x94\x4a\x33\xab\x8d\x75\x19\x4c\x2e\xcf\x2f\x2f\x26\xf0\x02\x1c\x2a\x15\xad\x00\x67\x80\x72\x2b\x6a\x84\x59\x85\xce\xca\x9c\xd8\xb8\xda\x48\xed\x5e\x12\xf0\xe6\xb4\x15\xac\x46\x36\xec\x58\x01\xd0\x61\xe1\x5f\x99\x02\xef\x47\x28\xf0\xcf\x1c\x9d\x48\x57\xcd\x1c\xad\x46\x87\x5e\x39\x43\x19\x28\xa9\x5b\x11\x0c\x9d\x5d\xcb\x1c\x5f\xe5\xb9\x69\xb4\xfb\x38\x3e\x01\x60\x6d\x54\x53\x61\xaf\x43\xd2\xea\xb0\x94\x2e\x59\xe1\xb6\x3f\x88\x18\x3e\x37\x1c\xdc\x3d\x19\xe4\x25\xbc\xa5\xf0\x91\x11\xad\x2a\x70\x21\x1a\xe5\x3e\x98\x02\x33\x38\xff\xf6\xfc\x1c\x5e\xc0\xa6\x44\x0d\x15\x6b\x83\x05\x58\x14\x45\x62\xb4\xda\x4e\x61\x83\xb0\x31\xfa\xa5\x83\x39\x82\x98\x2b\x64\x20\xf3\xb2\x32\xc5\x49\x2b\xf0\x05\x7c\x2a\x25\x81\x24\x10\xe0\xaa\x7a\x41\xd0\x10\x16\xb0\x30\x16\x96\xa8\xd1\x0a\x27\xf5\x12\xee\xef\x7f\x86\x15\x6e\x29\x85\x1b\x0d\xef\xff\x46\xf0\xe3\x15\x5c\xa4\x17\xe7\xd3\x5e\x4a\x77\x76\x30\x81\x40\x58\x8c\xf5\x20\xc3\xaa\x68\xc4\x02\x04\x10\xd6\x82\xa3\xa9\x05\x0a\x36\xd8\x8b\xc9\x85\x86\x8d\x95\x8e\x15\x4d\x0f\xe3\xb7\x44\xdd\x83\x81\x55\xed\xb6\x6f\xa4\x8d\x41\xac\xb0\x90\x4d\x95\xc1\x07\xac\x8c\xdd\xc6\x76\x22\x2c\x8c\x52\x66\xc3\x16\xb5\x47\x4b\xf2\xa6\x36\xc4\xcf\x04\xe4\x0d\x39\x53\x49\x46\x60\xa5\xcd\x46\xff\x5e\x1a\x72\xd4\x8b\x58\x48\x85\x53\xd8\x94\x32\x2f\x61\x6b\x1a\xd8\x48\xa5\x82\x51\xce\x40\x61\x38\x41\xf9\x31\x6f\xe2\xff\x58\x30\x1b\xcd\x6a\xf7\x02\x2c\xd6\x06\xac\x70\x25\x5a\x70\xa5\xd0\xed\xc1\x4b\xe9\xca\x66\x0e\x86\x1f\x22\x28\xb9\xc2\x14\x7e\x33\xcd\x4b\xa5\x40\x28\x32\xdd\x11\x63\xb0\x41\x3a\x90\xda\x19\xbf\x27\x37\xda\x09\xa9\xd1\x4e\x61\x8e\xca\x6c\x52\xb8\xc7\x01\xd5\xd2\xb9\x9a\xb2\xd9\xac\x30\x39\xa5\x1c\x58\x79\xc1\x61\x8d\x7a\xc6\x39\x4b\x6e\xb6\x6c\x64\x81\x34\x6b\x08\x93\xda\xca\xb5\x70\xe8\x43\x8f\x0d\x49\x4b\x57\xa9\x5e\x52\xe7\x0b\xa2\x32\xc9\x8d\x5e\xc8\x65\xff\x0a\x20\x3c\xf8\x20\xea\x2c\x7a\x18\x67\x60\x12\x6d\xfb\x5a\xbf\xf8\xd4\x9c\x05\x21\x43\xf8\x3d\xeb\x93\x8d\xa4\x92\x9f\x94\x62\x8d\x20\xa0\x90\x8b\x05\x5a\xae\xb6\x9d\x84\x36\xab\x86\x8a\xea\x5d\x10\xc4\xc5\x4e\xe0\xaa\xb4\x96\x05\x76\xb0\x2f\xe4\xb2\x12\xf5\xa0\x88\x74\x25\x08\x0d\xa8\x9d\xdd\x7a\x1b\x1e\xc2\xa2\x87\x29\x08\x5d\x40\xa3\x73\x53\x71\x99\xf7\xfb\x83\xb5\x1f\xbc\x3b\x85\x2e\x7a\x29\xa8\xd7\x5e\x82\x44\x6a\xfd\xb9\xe7\x01\x86\xe1\x2b\x3c\x10\x6d\x7b\xd6\x03\xbe\x12\x38\x03\xb2\xe2\x02\x0b\xd7\xb7\xd7\xbe\x08\xc0\x29\x9b\x45\x72\xa9\xa5\x1e\x0e\x67\xe3\xd6\x68\xe5\x42\xe6\xbe\xd2\x43\xdd\xd8\xda\x10\xd2\xd9\x17\x00\xd9\x4b\x09\xe5\x23\xa0\xc8\x00\xf1\x79\x5f\x00\x1c\x08\xbb\x1c\xd2\xf4\x09\xc4\x96\xf5\x92\xeb\x07\x45\xd0\x8c\x4b\xf0\x8b\x27\x8a\xf0\xfe\xbe\x03\x45\xb8\x83\xf3\xf1\x31\xf1\x8d\xfa\xb5\x07\xf9\x9d\x54\xf8\xda\x68\xc7\x7a\xfb\x86\xfd\x65\x88\x77\xd0\x88\xd6\xa3\x8d\x0d\x98\x32\xf0\xac\x53\x31\x92\x34\xec\x0f\xfd\xb8\x60\x05\x18\xa5\xa4\x4d\xb6\x84\x0b\x17\x30\x7e\x8c\x95\x92\x7a\x00\xdc\xd4\x5e\xb0\xa4\xee\xcc\x22\x85\x9b\x5e\xfd\x57\xf4\xba\x8b\x28\x5e\x42\xe8\x58\x3b\x67\x1b\x9c\x8e\xdc\x36\xd6\x92\x9b\x09\x51\x53\x05\x5b\xb8\xf3\x40\x2f\x66\x0a\x86\xab\xdf\x46\x12\xfa\x02\xd6\x2f\xed\x65\x75\x5b\xee\xbd\x1f\xd2\x83\xb0\xc6\x7a\xf5\xb0\x26\x71\x9c\xb7\x86\x07\x68\xfa\xee\xb0\x97\x1f\x3b\xd9\x31\x4a\x28\x3e\x92\x19\x54\xd2\x9f\xd0\x2a\x71\xaa\x8d\x3b\xa0\xc9\xd9\xff\xa4\xca\x17\x74\xff\x2f\xd0\xa6\x65\x78\x27\xbd\x79\xa1\x0b\xec\x71\x8f\x88\x9d\xb4\x51\x63\xd1\xf7\x68\x6d\x60\x92\x85\x2e\x30\x01\x59\x89\x25\x52\x1f\x65\x29\xbc\x93\xba\xf0\x91\x54\x71\x4b\xb3\x98\x0f\x15\x33\xb4\x33\x85\x82\x90\x1b\x97\xdf\xca\x05\x80\xc9\x2d\x08\xd7\xf7\x9c\xb2\x99\xa7\x85\xc9\x57\x68\xd3\xdc\x54\x33\x3b\x0b\xfd\xc7\xff\x9a\x39\xd1\xa7\x6d\x57\x43\x98\xa4\x32\x81\xe5\x53\x9d\x58\x02\x6b\x9a\xf6\x6b\xfc\x31\x19\xb4\x02\xa5\x89\xa5\x65\x17\xe9\xc5\x5f\xd3\x8b\xf1\xda\xdb\x46\xa9\x5b\xa3\x64\xbe\xcd\xe0\x66\xf1\xd1\xb8\x5b\x8b\x14\x5b\x61\x91\x4c\x63\x73\xa4\xd8\x15\x16\xff\xd3\x20\xb9\xd1\x33\x80\xbc\x6e\x32\xf8\xee\xbc\x1a\x3d\xac\x3c\xcd\xc8\xe0\xfb\x6f\x3f\xc8\x81\xdb\x1a\x1b\x6f\x4e\x06\xcf\xdc\x7a\x9e\x7b\x79\x7e\xc9\xac\x4d\xea\x85\xb1\x95\x4f\x1a\xa1\xfa\xd5\x4a\xae\x51\x23\xd1\xad\x35\x73\x8c\x35\x60\x48\xaf\xc7\x31\x13\x8e\x0a\x02\xc7\x8f\x85\x2b\x33\x98\x89\x5a\x06\xa4\xd7\xdf\xcf\x64\x81\xda\x49\xb7\x4d\xeb\x66\x1e\xad\x95\x5a\x3a\x29\xd4\x1b\x54\x62\x7b\xcf\xbd\xa1\xa0\x0c\xbe\x8b\x16\x38\x59\xa1\x69\xdc\x81\x77\x4c\xf0\xe4\xff\x87\xaa\x51\xc3\x18\x39\xe6\x30\x35\x87\x40\xb1\x6e\x83\x66\xe8\x72\xaf\x59\x31\x23\x2a\x43\x85\xf2\xd7\x25\x50\xa6\xed\x75\x4b\x76\x19\x48\x1d\x62\xee\x25\x85\x3d\x44\xe5\x6c\x94\xc3\x1d\x66\xbf\x6a\xb5\xcd\x7c\x19\x65\x69\xcc\xbf\x7d\x77\x9c\xb7\xa4\x82\x53\xaa\x46\xbb\x30\x36\x47\x16\x1a\x08\x37\xf3\xed\xa7\x14\x8f\x39\xf1\x58\xf7\xb5\xb0\xad\xee\x61\xd9\xd7\xa9\x1f\xe5\xe8\x8d\xce\x55\xe3\xbb\x36\x5f\x1b\x02\xb9\xea\x3a\x7a\xe0\xa5\xcf\xd0\xe8\x8e\x48\xff\xc0\x5b\x77\x28\xee\xd0\xcc\x0a\xcc\x95\xb0\x7c\x5d\x98\x9b\x75\x54\x00\x8e\x50\xd0\xd0\x9a\x63\xe3\xad\x31\x6e\x96\x12\x95\x4f\x1a\x20\xf4\xe8\xd4\xc9\x40\x8f\x26\xe1\xe4\x69\xb7\x24\x92\x80\x7a\x2d\xad\xd1\x9e\x8c\x04\x9e\x37\x79\xff\xf9\xa7\xb7\xaf\x7f\xfd\xf8\xee\xe6\x7a\x12\xe8\xc7\x94\xf1\x30\x6b\xb4\x76\xcc\x15\x23\x31\xbe\xd9\xcf\xb7\x81\xc9\x39\x75\xc8\xc6\x3d\x92\xb7\x6f\xe3\x10\x9c\xbc\xf8\x49\x43\x99\x54\xf0\x6d\xb9\x3b\x8d\x4b\x74\x44\x83\x5b\xed\xbc\x4f\x22\x11\xbb\x64\x3a\x76\xba\x67\xd2\xdd\xb5\x4f\x68\x10\xca\xa1\xd5\x7c\xad\xdb\xd3\x78\x61\x4d\xd5\x73\x99\x8a\x49\x80\x20\x0e\xb7\x96\xd1\x31\x0c\xca\xe4\x2b\xda\x77\x36\xea\x75\x76\x00\x97\x01\xee\x11\x2e\x6b\xa1\x1a\xdc\xc3\xe4\xb9\x20\xde\x8d\x81\x8e\xef\x1d\x89\x00\xa6\x9b\x63\x9a\x79\x84\x68\x3e\x11\x97\xbc\x2a\x30\xeb\xd1\xba\x71\x7d\x18\x94\x6e\xe7\x3f\xfb\xb4\x72\x20\x04\xfd\xd5\x90\xa2\x80\xeb\xcc\x0a\x3e\x60\x83\x0e\xdc\x1a\xf7\xea\xca\x11\xe2\xf2\x44\xf8\x45\x8a\xee\xd2\x92\xa7\xca\xc6\x46\x68\xcf\x27\xa9\xa9\x6b\xb5\x85\x9f\x3f\x7d\xba\x85\xb9\x20\x99\x83\x68\x5c\xc9\x64\xd6\xb7\x01\xa1\x02\x25\x19\x2e\xd2\x2c\x70\x2d\x85\xb7\xe1\xe1\xfa\xe6\xd3\xef\xaf\x3e\x7f\xfa\xf9\xf3\xfd\xdb\xbb\x07\xef\xab\xfe\xd1\xfb\xb7\xbf\x3d\x8c\xb2\x75\x2d\xac\x14\x73\x85\xd4\xdd\x2c\x23\x81\x34\xe2\x9b\x6d\xf0\xbd\xb3\xa6\x1a\x07\x60\x58\x76\x87\x8b\x6c\xe4\xb6\x11\x8d\xe4\xaa\xcc\x26\x0c\x00\x70\xc0\x64\x23\x3c\x02\x04\x61\xb8\x83\x05\xd3\x88\x5c\xe4\x25\x73\x77\x3d\x4a\xcc\xfe\x3e\xca\x48\xb1\xf4\x69\x24\xc5\xd8\xf6\xc2\x19\x6d\x68\x87\x53\x7e\xe3\xd4\x1f\x92\x0b\xdd\x61\xec\x4a\xa4\x38\x90\x87\x4b\x88\xdb\x18\xd6\xb2\x61\x9c\x7c\xb9\xf0\x23\x38\xef\x5a\x28\xcd\xc6\x0f\x8e\x8c\xd6\x98\x7b\x97\x49\x37\x0e\xfc\x24\xe9\x0d\xf0\x53\x03\x3e\xfc\xaa\x7f\x94\xb6\xb7\xa5\x94\xd6\x79\x9a\xab\x86\x1c\xda\x94\xbb\x8f\x8a\x21\xf9\x4c\xa1\x50\x0e\x50\xbc\x0e\x4b\x6f\x6e\x47\x46\x71\xcd\x24\x74\x7e\x30\x35\x4e\xcb\x41\x87\x6e\x3d\x47\x97\xb3\xbc\xd2\x8f\x8a\xa2\x38\x8f\x35\x6e\x57\x5f\x9d\x8c\x28\xb2\x24\xa8\x1a\xf2\xa3\x33\x8f\x9e\xc4\x22\xd4\x82\xb9\xef\xca\x9e\xa0\xfa\x89\xd9\x69\x37\x86\x3a\xdb\x4d\xc4\xae\x7f\x4d\x7d\x00\x47\x83\xb3\x91\x22\xdc\xc9\x42\x77\x4e\x0a\x69\xaf\xf6\x7a\x76\xac\xd6\x5d\x44\x8f\x07\xe7\x7d\xbe\xfb\x25\x4c\xf6\x84\x5e\x86\x77\xd7\xd2\xf9\x69\x13\x49\x67\xec\xb6\xef\x35\xef\x98\xd6\x47\xe2\x8e\xe5\x1c\x87\x4d\x64\x7b\x9b\x32\x07\xd3\x29\xce\x85\x8e\xf8\xff\xe5\x34\xce\xcc\xb3\x6c\xf8\xfb\xfd\xdb\xdf\xce\xfe\x11\x66\x5e\xfe\x4e\xd0\x10\xda\xd9\xa0\x6c\x1a\x27\x7a\x77\xef\xbb\x96\x8e\x8d\x8c\x6a\x1d\x23\xc7\x89\xd6\x58\x75\xf5\xf8\x78\x68\x85\xaf\x44\x8a\x10\x46\x15\x32\xde\xb7\x94\x2e\x52\x24\x8b\x2e\x13\xc9\x12\x5d\x42\x4e\x58\x87\x07\x4b\xdb\xae\x16\x73\x2b\x74\x5e\x76\x8a\xfc\xe4\xff\xda\xd5\xa5\xb5\x83\x2b\x27\x1d\x92\xc1\x44\x98\x25\xdc\xfb\x88\xa5\x7f\x1a\xa9\xa3\x0d\x93\xe9\xe4\x79\xe3\x8e\x95\x59\x8b\x9c\x0a\x79\x18\xa0\x54\x42\xcb\x05\x5f\x71\x38\xab\x49\x16\x68\x43\x80\xec\x5c\x14\x7d\xa3\x30\x84\xd0\xe8\x02\xed\x4e\xd4\x59\x54\xc2\xc9\x35\x7a\x06\x4f\x5d\x4e\x2c\x47\x91\x77\x00\x78\x6f\x26\x35\xf3\x42\xda\x8b\x69\xf8\xfd\xcd\x31\x88\x23\xe8\xfc\x07\x8a\x43\xd0\xf9\xa9\x7f\x87\xfe\xde\xaa\xe3\x42\x3f\x13\xda\x83\x81\x45\x68\xfb\xc8\xda\x59\x73\x5c\xe2\xdb\x4a\xc8\x83\x6a\x22\xbf\xe8\x64\xee\xad\x7a\xae\x73\x22\x17\xc8\x8d\x61\xa7\xa0\xf6\xd3\x7c\xc6\x9a\x49\x94\x74\x3b\xf3\xb8\x18\xef\x96\x8e\xb4\x64\xe3\xea\x08\xfb\xe8\x76\xb4\xb2\x78\xd7\xd5\xdf\x57\xb8\x05\x59\xfc\x78\x8c\x59\x46\xda\xf0\x56\xe1\x1a\x8b\xa3\x61\xe0\x81\x33\xfc\xeb\x6d\xd2\xaf\xa7\x51\xf1\xed\x7a\x0f\x48\x07\xa5\x20\xcf\x8a\x8c\x56\x5b\x10\x79\x8e\x14\xfa\x53\x89\x61\x9e\x7e\xda\x8d\x6e\x1f\x16\x42\x11\x3e\x9c\x9d\x44\xae\xb8\x6b\xe9\xd4\x21\x6f\x74\x42\x4f\x8e\x17\x8a\xf1\xb2\x03\xfe\x21\x67\x9b\xdc\x05\x7d\x37\x7e\xb2\xc2\x6c\xbb\x71\x40\x5b\x9d\xc3\xdc\x98\xd5\x0a\xb1\xe6\x8c\xe9\x55\x9d\x2c\xa5\x9b\x4c\xa1\x42\xc1\x40\x73\x79\x06\xe1\xc7\x1d\x6d\x12\x35\x35\x39\x8b\xa2\xea\xb3\xe9\x6c\x47\x31\x16\xcd\xf5\xc9\x21\x17\xb1\xa7\xe3\x45\xe3\x1f\xae\x0b\x9a\xa8\x7f\x0b\x0d\x93\xee\x8c\x49\xd7\x5d\x23\x21\xa7\x98\x2e\xd3\x29\xfc\x1b\x99\xe4\xbf\x56\xa6\x29\xce\x52\x3f\x6b\x74\x66\x85\x61\x5c\x28\xac\x93\x79\xa3\x84\xed\x9c\xd1\x4a\xd9\x25\x06\xed\xa9\x57\x1b\xe2\xae\x90\xb3\xac\x74\xc3\x72\xd3\x8d\xb1\x2b\xea\xef\xfd\x3b\xdb\xfc\x41\x57\x62\x9e\x5f\x7c\x73\xb9\xff\x6f\x6c\xf0\xdb\x10\x7d\x5d\x45\xeb\xbf\x5b\x19\x7d\x24\x34\x3e\xb4\xab\xaf\x87\xc5\x3b\x11\xd2\xc9\x4b\x06\x79\x57\x9e\x92\x3f\x1d\x2d\x87\xb6\xf8\x83\x9f\x08\x9d\x7b\xb4\xeb\x03\x5f\x34\xf9\x6e\x36\xf0\x39\xce\xd1\x1f\x62\x62\x21\x56\xdc\x94\x43\x94\x11\xba\xe8\x33\xe9\xcb\xe8\x4b\x6b\xf4\xc9\x94\x9d\xe3\x27\xf8\xfe\x7e\x94\x8e\xac\x54\x92\x1c\xea\xa4\x55\xe1\x2a\xbb\x3c\xbf\xbc\x88\x55\x7c\x23\xc9\xa3\x4b\xb9\xd0\x3e\x54\xcd\x02\x2c\x2e\x25\x85\x8f\x13\xa7\xdd\xe5\x29\xcc\xff\x0a\xe1\xc4\x5c\x10\x0e\x38\xdf\x85\xb5\xdb\x56\xce\x7d\x27\x66\x07\xec\x56\xe4\x36\x29\xc2\xba\xa4\x3b\xef\x29\xb4\x9f\xdb\x11\xb0\x6e\x2f\x4d\xaf\x8a\x42\x86\x21\x1b\xf3\x9f\x57\x7c\x79\x1b\x49\x1b\xde\x0f\x14\x78\x54\x95\xad\xe7\x55\xcf\xc8\x49\xfc\x27\xf8\x27\x0a\xfa\xf0\x35\xbe\x3b\xf4\xd7\xba\x3d\xf2\xcd\xc7\xfb\x8e\xcf\xd2\xb4\x1f\xa5\xb7\xec\x16\x74\x61\x1c\x75\x63\xfa\x4a\x6c\xfd\xc0\x52\xad\x87\xd9\xbb\x26\x65\xcc\xaa\xa9\x41\x12\x35\x48\x60\x34\x90\xa9\x10\xde\xf7\x9f\xb0\x59\x7a\x53\xd3\xf0\x45\xa4\xd0\xd4\xcd\x44\x27\x1f\x8d\xc6\x49\xfc\x26\xdc\x30\xe3\x6f\x22\xe1\x70\x1a\x7f\x26\xe9\x6e\x8a\x5e\xbf\xd1\x9b\xfe\x1e\x3e\xb9\x98\x9c\xfc\x37\x00\x00\xff\xff\x3d\xc9\xd9\x97\xf6\x20\x00\x00"),
 		},
 		"/flux-secret.yaml.tmpl": &vfsgen۰CompressedFileInfo{
 			name:             "flux-secret.yaml.tmpl",
@@ -65,6 +72,7 @@ var templates = func() http.FileSystem {
 		},
 	}
 	fs["/"].(*vfsgen۰DirInfo).entries = []os.FileInfo{
+		fs["/#flux-deployment.yaml.tmpl#"].(os.FileInfo),
 		fs["/flux-account.yaml.tmpl"].(os.FileInfo),
 		fs["/flux-config.yaml.tmpl"].(os.FileInfo),
 		fs["/flux-deployment.yaml.tmpl"].(os.FileInfo),
